@@ -19,56 +19,52 @@
     "@apollo/client": "^3.3.21",
     "@apollo/react-hooks": "^4.0.0",
     "@arbitrum/sdk": "^2.0.0",
-    "@craco/craco": "^6.3.0",
+    "@craco/craco": "^7.1.0",
     "@eth-optimism/sdk": "^1.0.0",
     "@ethersproject/contracts": "^5.6.2",
     "@ethersproject/networks": "^5.6.4",
     "@gitcoinco/passport-sdk-reader": "^0.1.4",
-    "@portis/web3": "^4.0.5",
     "@ramp-network/ramp-instant-sdk": "^2.2.0",
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.8",
     "@uniswap/sdk": "^3.0.3",
     "@uniswap/v2-periphery": "^1.1.0-beta.0",
-    "@walletconnect/web3-provider": "^1.5.2",
+    "@web3modal/ethers5": "^3.5.3",
     "antd": "4.16.0",
     "apollo-boost": "^0.4.9",
     "apollo-client": "^2.6.10",
     "apollo-utilities": "^1.3.4",
-    "authereum": "^0.1.14",
     "axios": "^0.25.0",
     "bl": "^5.0.0",
     "bnc-notify": "^1.5.0",
     "dotenv": "^8.2.0",
     "eslint-plugin-prettier": "^4.0.0",
-    "eth-hooks": "2.3.19",
-    "ethers": "^5.4.1",
-    "fortmatic": "^2.2.1",
-    "graphiql": "^1.4.7",
+    "eth-hooks": "4.6.2",
+    "ethers": "^5.7.2",
     "graphql": "^15.3.0",
     "ipfs-http-client": "^55.0.0",
     "isomorphic-fetch": "^3.0.0",
     "moment": "^2.29.4",
     "node-watch": "^0.7.1",
     "qrcode.react": "^1.0.0",
-    "react": "^17.0.2",
+    "react": "18.2.0",
     "react-blockies": "^1.4.1",
     "react-css-theme-switcher": "^0.2.2",
-    "react-dom": "^17.0.2",
+    "react-dom": "18.2.0",
     "react-loading": "^2.0.3",
     "react-qr-reader": "^2.2.1",
-    "react-router-dom": "^6.3.0",
-    "react-scripts": "4.0.3",
+    "react-query": "^3.39.2",
+    "react-router-dom": "^6.21.1",
+    "react-scripts": "5.0.1",
     "searchico": "^1.2.1",
-    "typescript": "^5.1.3",
-    "walletlink": "^2.1.5",
-    "web3modal": "^1.9.1"
+    "typescript": "^5.1.3"
   },
   "devDependencies": {
     "@apollo/client": "^3.3.21",
     "@testing-library/dom": "^6.12.2",
-    "@types/react": "^16.9.19",
+    "@types/react": "^18.2.0",
+    "@types/react-dom": "^18.2.0",
     "autoprefixer": "^9",
     "chalk": "^4.1.0",
     "eslint-config-prettier": "^8.3.0",
