@@ -1,20 +1,17 @@
 import React from "react";
 
-export default function Account({ web3Modal, loadWeb3Modal }) {
+import { useWeb3Modal, useWeb3ModalAccount } from "@web3modal/ethers5/react";
+
+export default function Account() {
+  const { isConnected } = useWeb3ModalAccount();
+  const { open } = useWeb3Modal();
   return (
     <div>
-      {web3Modal?.cachedProvider ? (
-        <></>
-      ) : (
+      {!isConnected && (
         <div className="flex">
-          {web3Modal && (
-            <button
-              className="rounded-sm bg-purple-connectPurple py-4 px-10 text-white text-base"
-              onClick={loadWeb3Modal}
-            >
-              Connect Wallet
-            </button>
-          )}
+          <button className="rounded-sm bg-purple-connectPurple py-4 px-10 text-white text-base" onClick={open}>
+            Connect Wallet
+          </button>
         </div>
       )}
     </div>
